requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# API clients
shodan>=1.31.0

# Docker
docker>=7.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Async support
aiohttp>=3.9.0
